Web app/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
Flask-WTF==1.2.1
WTForms==3.1.1
cryptography==41.0.7
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
email-validator==2.1.0
Werkzeug==3.0.1
itsdangerous==2.1.2
pyotp==2.9.0
qrcode==7.4.2
bleach==6.1.0
MarkupSafe==2.1.3
python-dateutil==2.8.2
gunicorn==21.2.0
